requirements.txt
matplotlib==3.10.1
numpy==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
seaborn==0.13.2
streamlit==1.42.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
tqdm==4.67.1
